requirements.txt
discord.py
aiosqlite
watchdog
fastapi
uvicorn
sqlite-web
Pillow
